Split up the files into separate files
