+++ packages/react-frontend/src/pages/game/VoteMovingBox.jsx
@@ -1,590 +1,41 @@
-import React, { useCallback, useEffect, useMemo, useRef, useState } from "react";
-import { useLocation, useNavigate } from "react-router-dom";
+import { useCallback, useEffect, useMemo, useRef, useState } from "react";
 import "./VoteMovingBox.css";
-import UserCrabIcon from "../../assets/user-crab.png";
-import CrownIcon from "../../assets/hats/crown.png";
 import { authFetch } from "../../authFetch";
 import frontendLink from "../../frontendLink";
+import { readAccountPlaylist } from "../Playlist/playlistStorage";
+import CrabLane from "./CrabLane";
+import CurrentSongPlayer from "./CurrentSongPlayer";
+import SongPicker from "./SongPicker";
+import VoteMovingBoxItem from "./VoteMovingBoxItem";
 import {
-  getSongArtist,
-  getSongId,
-  getSongLink,
-  getSongVideoId,
-  getSongTitle,
-  readAccountPlaylist,
-} from "../Playlist/playlistStorage";
+  MAX_SCORE,
+  MIN_SCORE,
+  ROUND_SECONDS,
+  STACK_CARD_GAP,
+  STACK_CARD_HEIGHT,
+  VOTE_ARENA_CHROME_HEIGHT,
+  VOTE_ARENA_MAX_HEIGHT,
+  VOTE_ARENA_MIN_HEIGHT,
+  VOTE_PLAYBACK_ARENA_HEIGHT,
+} from "./constants";
 import {
-  createCrabIcon,
-  getHatSourceForCrab,
-  normalizeCrabProfile,
-  readStoredCrabProfile,
-} from "../profile/crabColor";
-
-const hatImages = import.meta.glob("../../assets/hats/*.png", {
-  eager: true,
-  query: "?url",
-  import: "default",
-});
-
-const DEFAULT_USERS = [
-  { id: "captain", name: "Captain" },
-  { id: "sailor", name: "Sailor" },
-  { id: "dj", name: "DJ" },
-  { id: "guest", name: "Guest" },
-];
+  clamp,
+  formatTime,
+  getNextQueueColorIndex,
+  getQueueAfterWinnerStarts,
+  getQueueEntryScore,
+  getRoomTimeLeft,
+  getWinningQueueEntry,
+  isSameQueueEntry,
+  makeCurrentSongFromQueueEntry,
+  makeLocalEntryId,
+  normalizeCurrentSong,
+  normalizeQueueEntries,
+  normalizeSongs,
+  normalizeUsers,
+} from "./gameUtils";
 
 const API = frontendLink;
-const ROUND_SECONDS = 120;
-const MIN_SCORE = -20;
-const MAX_SCORE = 40;
-const STACK_CARD_HEIGHT = 86;
-const STACK_CARD_GAP = 30;
-const VOTE_ARENA_MIN_HEIGHT = 500;
-const VOTE_ARENA_MAX_HEIGHT = 745;
-const VOTE_ARENA_CHROME_HEIGHT = 132;
-const VOTE_PLAYBACK_ARENA_HEIGHT = 640;
-const VOTE_BOX_COLORS = [
-  { start: "#fff4cc", middle: "#ffd67a", end: "#f2a84a" },
-  { start: "#d9f7ff", middle: "#8edff2", end: "#4db4d7" },
-  { start: "#e7f7d5", middle: "#addf76", end: "#73b95c" },
-  { start: "#ffe2ec", middle: "#f5a0bd", end: "#d86f9b" },
-  { start: "#e9e1ff", middle: "#bca3ee", end: "#8b75d6" },
-  { start: "#ffe8d3", middle: "#f5b174", end: "#dc7e45" },
-  { start: "#daf5ea", middle: "#8ed6ba", end: "#4aa889" },
-  { start: "#e3edff", middle: "#9ebdf1", end: "#638bd8" },
-  { start: "#f7e6bd", middle: "#e2c464", end: "#b99b3b" },
-  { start: "#f0dcff", middle: "#d09cf1", end: "#aa6ed0" },
-];
-
-let youtubeApiPromise;
-
-function clamp(value, min, max) {
-  return Math.min(Math.max(value, min), max);
-}
-
-function loadYouTubeApi() {
-  if (typeof window === "undefined") {
-    return Promise.resolve(null);
-  }
-
-  if (window.YT?.Player) {
-    return Promise.resolve(window.YT);
-  }
-
-  if (!youtubeApiPromise) {
-    youtubeApiPromise = new Promise((resolve) => {
-      const previousReady = window.onYouTubeIframeAPIReady;
-
-      window.onYouTubeIframeAPIReady = () => {
-        previousReady?.();
-        resolve(window.YT);
-      };
-
-      if (!document.querySelector('script[src="https://www.youtube.com/iframe_api"]')) {
-        const script = document.createElement("script");
-        script.src = "https://www.youtube.com/iframe_api";
-        document.body.appendChild(script);
-      }
-    });
-  }
-
-  return youtubeApiPromise;
-}
-
-function getPlayableVideoId(song) {
-  return getSongVideoId(song) || getSongVideoId({ songLink: song?.songLink });
-}
-
-function normalizeUsers(users, currentUserName) {
-  const sourceUsers = users?.length ? users : DEFAULT_USERS;
-
-  return sourceUsers.map((user, index) => {
-    const name =
-      typeof user === "string" ? user : user.name || user.userName || `User ${index + 1}`;
-    const savedCrab =
-      typeof user === "object" && user?.crab
-        ? user.crab
-        : name === currentUserName
-          ? readStoredCrabProfile()
-          : {};
-
-    return {
-      id: user.id || user.userId || name,
-      name,
-      crab: normalizeCrabProfile(savedCrab),
-    };
-  });
-}
-
-function normalizeSongs(songs) {
-  const sourceSongs = Array.isArray(songs) ? songs : [];
-
-  return sourceSongs.map((song, index) => ({
-    id: getSongId(song) || `${getSongTitle(song)}-${index}`,
-    name: getSongTitle(song),
-    artist: getSongArtist(song) || "Unknown artist",
-    songLink: getSongLink(song),
-    videoId: getSongVideoId(song),
-    source: song.source || "My Playlist",
-  }));
-}
-
-function formatTime(seconds) {
-  const minutes = Math.floor(seconds / 60);
-  const remainingSeconds = String(seconds % 60).padStart(2, "0");
-
-  return `${minutes}:${remainingSeconds}`;
-}
-
-function makeLocalEntryId() {
-  return `entry-${Date.now()}-${Math.random().toString(36).slice(2, 7)}`;
-}
-
-function getQueueEntryScore(entry) {
-  return Number.isFinite(entry?.score) ? entry.score : entry?.upvotes || 0;
-}
-
-function getStableColorIndex(value) {
-  const text = String(value || "");
-  let hash = 0;
-
-  for (let index = 0; index < text.length; index += 1) {
-    hash = (hash * 31 + text.charCodeAt(index)) % VOTE_BOX_COLORS.length;
-  }
-
-  return hash;
-}
-
-function getNextQueueColorIndex(queue = []) {
-  const usedColorIndexes = new Set(
-    queue
-      .map((entry) => entry.colorIndex)
-      .filter((colorIndex) => Number.isInteger(colorIndex))
-  );
-
-  for (let colorIndex = 0; colorIndex < VOTE_BOX_COLORS.length; colorIndex += 1) {
-    if (!usedColorIndexes.has(colorIndex)) {
-      return colorIndex;
-    }
-  }
-
-  return queue.length % VOTE_BOX_COLORS.length;
-}
-
-function normalizeQueueEntries(queue) {
-  const sourceQueue = Array.isArray(queue) ? queue : [];
-
-  return sourceQueue.map((entry, index) => {
-    const songId = entry.songId || entry.id || `${entry.name || "song"}-${index}`;
-    const entryId = entry.entryId || songId;
-
-    return {
-      entryId,
-      song: {
-        id: songId,
-        name: entry.name || entry.title || "Untitled song",
-        artist: entry.artist || "Unknown artist",
-        songLink: entry.songLink || "",
-        videoId: entry.videoId || "",
-        source: entry.source || "Room queue",
-      },
-      score: getQueueEntryScore(entry),
-      colorIndex: Number.isInteger(entry.colorIndex)
-        ? entry.colorIndex
-        : getStableColorIndex(entryId),
-      addedBy: entry.addedBy || null,
-    };
-  });
-}
-
-function normalizeCurrentSong(song) {
-  if (!song) {
-    return null;
-  }
-
-  if (typeof song === "string") {
-    return { name: song, artist: "" };
-  }
-
-  return {
-    entryId: song.entryId || "",
-    songId: song.songId || song.id || "",
-    name: song.name || song.title || "Untitled song",
-    artist: song.artist || "",
-    songLink: song.songLink || "",
-    videoId: song.videoId || "",
-    score: getQueueEntryScore(song),
-  };
-}
-
-function getWinningQueueEntry(queue) {
-  const entries = Array.isArray(queue) ? queue : [];
-  if (!entries.length) return null;
-
-  return entries.reduce((leader, entry) =>
-    getQueueEntryScore(entry) > getQueueEntryScore(leader) ? entry : leader
-  );
-}
-
-function makeCurrentSongFromQueueEntry(entry) {
-  return {
-    entryId: entry.entryId || entry.songId,
-    songId: entry.songId,
-    name: entry.name,
-    artist: entry.artist,
-    songLink: entry.songLink || "",
-    videoId: entry.videoId || "",
-    score: getQueueEntryScore(entry),
-  };
-}
-
-function isSameQueueEntry(entry, targetEntry) {
-  if (!targetEntry) return false;
-
-  return (entry.entryId || entry.songId) === (targetEntry.entryId || targetEntry.songId);
-}
-
-function getQueueAfterWinnerStarts(queue, winningEntry, options = {}) {
-  const sourceQueue = Array.isArray(queue) ? queue : [];
-
-  if (options.continuousPlaylistMode === "playQueue") {
-    return sourceQueue.filter((entry) => !isSameQueueEntry(entry, winningEntry));
-  }
-
-  return [];
-}
-
-function getRoomTimeLeft(room, now) {
-  const roundSeconds = room?.roundSeconds ?? ROUND_SECONDS;
-
-  if (room?.timerPaused) {
-    return clamp(room.timerRemainingSeconds ?? roundSeconds, 0, roundSeconds);
-  }
-
-  if (now == null) {
-    return clamp(room?.timerRemainingSeconds ?? roundSeconds, 0, roundSeconds);
-  }
-
-  if (room?.roundEndsAt) {
-    return clamp(Math.ceil((new Date(room.roundEndsAt).getTime() - now) / 1000), 0, roundSeconds);
-  }
-
-  return clamp(room?.timerRemainingSeconds ?? roundSeconds, 0, roundSeconds);
-}
-
-function VoteMovingBoxItem({ canDelete, entry, isVotingPaused, onDelete, onVote, stackIndex }) {
-  const { song, score } = entry;
-  const color = VOTE_BOX_COLORS[entry.colorIndex % VOTE_BOX_COLORS.length];
-
-  return (
-    <article
-      className="vote-box-row"
-      style={{
-        "--vote-box-start": color.start,
-        "--vote-box-middle": color.middle,
-        "--vote-box-end": color.end,
-        transform: `translateY(${stackIndex * (STACK_CARD_HEIGHT + STACK_CARD_GAP)}px)`,
-      }}
-    >
-      <div className="vote-box-bar" aria-label={`${song.name} score ${score}`}>
-        <span className="vote-box-score">{score}</span>
-        <div className="vote-box-track">
-          <span className="vote-box-track-title">{song.name}</span>
-          <span className="vote-box-track-artist">{song.artist}</span>
-        </div>
-        <div className="vote-box-actions">
-          <button
-            type="button"
-            className="vote-box-button vote-box-button--up"
-            onClick={() => onVote(entry.entryId, 1)}
-            disabled={isVotingPaused}
-            aria-label={`Upvote ${song.name}`}
-          >
-            Up
-          </button>
-          <button
-            type="button"
-            className="vote-box-button vote-box-button--down"
-            onClick={() => onVote(entry.entryId, -1)}
-            disabled={isVotingPaused}
-            aria-label={`Downvote ${song.name}`}
-          >
-            Down
-          </button>
-          {canDelete && (
-            <button
-              type="button"
-              className="vote-box-button vote-box-button--delete"
-              onClick={() => onDelete(entry.entryId)}
-              aria-label={`Delete ${song.name}`}
-            >
-              Delete
-            </button>
-          )}
-        </div>
-      </div>
-    </article>
-  );
-}
-
-function YouTubeSongPlayer({ onEnded, song }) {
-  const playerHostRef = useRef(null);
-  const playerRef = useRef(null);
-  const endedRef = useRef(false);
-  const onEndedRef = useRef(onEnded);
-  const bufferingTimeoutRef = useRef(null);
-  const videoId = getPlayableVideoId(song);
-  const [playerError, setPlayerError] = useState({ videoId: "", message: "" });
-  const activePlayerError = playerError.videoId === videoId ? playerError.message : "";
-
-  useEffect(() => {
-    onEndedRef.current = onEnded;
-  }, [onEnded]);
-
-  useEffect(() => {
-    endedRef.current = false;
-
-    function clearBufferingTimeout() {
-      clearTimeout(bufferingTimeoutRef.current);
-      bufferingTimeoutRef.current = null;
-    }
-
-    function startBufferingTimeout() {
-      clearBufferingTimeout();
-      bufferingTimeoutRef.current = setTimeout(() => {
-        if (isActive) {
-          setPlayerError({ videoId, message: "This song got stuck loading." });
-        }
-      }, 10000);
-    }
-
-    if (!videoId) {
-      return undefined;
-    }
-
-    let isActive = true;
-    const timeoutId = setTimeout(() => {
-      if (isActive && !playerRef.current) {
-        setPlayerError({ videoId, message: "This song is taking too long to load." });
-      }
-    }, 8000);
-
-    loadYouTubeApi().then((YT) => {
-      if (!isActive || !YT?.Player || !playerHostRef.current) {
-        if (isActive) {
-          setPlayerError({ videoId, message: "The YouTube player could not load." });
-        }
-        return;
-      }
-
-      playerRef.current = new YT.Player(playerHostRef.current, {
-        videoId,
-        playerVars: {
-          autoplay: 1,
-          controls: 1,
-          modestbranding: 1,
-          playsinline: 1,
-          rel: 0,
-        },
-        events: {
-          onReady() {
-            clearTimeout(timeoutId);
-          },
-          onError() {
-            if (isActive) {
-              clearBufferingTimeout();
-              setPlayerError({ videoId, message: "This YouTube song could not be played here." });
-            }
-          },
-          onStateChange(event) {
-            if (event.data === YT.PlayerState.PLAYING) {
-              clearTimeout(timeoutId);
-              clearBufferingTimeout();
-            }
-
-            if (event.data === YT.PlayerState.BUFFERING) {
-              startBufferingTimeout();
-            }
-
-            if (event.data === YT.PlayerState.ENDED && !endedRef.current) {
-              clearBufferingTimeout();
-              endedRef.current = true;
-              onEndedRef.current?.();
-            }
-          },
-        },
-      });
-    });
-
-    return () => {
-      isActive = false;
-      clearTimeout(timeoutId);
-      clearBufferingTimeout();
-      playerRef.current?.destroy?.();
-      playerRef.current = null;
-    };
-  }, [song?.entryId, videoId]);
-
-  if (!videoId || activePlayerError) {
-    return (
-      <div className="vote-song-player-missing">
-        <p>
-          {activePlayerError || "This winning song does not have a playable YouTube link saved."}
-        </p>
-        {song?.songLink ? (
-          <a href={song.songLink} target="_blank" rel="noreferrer">
-            Open song
-          </a>
-        ) : null}
-      </div>
-    );
-  }
-
-  return <div className="vote-song-player-frame" ref={playerHostRef} />;
-}
-
-function CurrentSongPlayer({ canControl, onComplete, shouldPlayAudio, song }) {
-  if (!shouldPlayAudio) {
-    return (
-      <section className="vote-current-song" aria-label="Current song">
-        <div className="vote-song-player-missing">
-          <p>The host is playing this song on their computer.</p>
-        </div>
-      </section>
-    );
-  }
-
-  return (
-    <section className="vote-current-song" aria-label="Current song">
-      <YouTubeSongPlayer song={song} onEnded={canControl ? onComplete : undefined} />
-    </section>
-  );
-}
-
-function SongPicker({ isGuest, onAddSong, onLoginRequired, songs }) {
-  const location = useLocation();
-  const navigate = useNavigate();
-  const [searchTerm, setSearchTerm] = useState("");
-  const filteredSongs = useMemo(() => {
-    const normalizedSearch = searchTerm.trim().toLowerCase();
-
-    if (!normalizedSearch) {
-      return songs;
-    }
-
-    return songs.filter((song) => song.name.toLowerCase().includes(normalizedSearch));
-  }, [songs, searchTerm]);
-  const handleOpenPlaylist = () => {
-    if (isGuest) {
-      onLoginRequired?.();
-      return;
-    }
-
-    navigate("/home/playlist", {
-      state: {
-        returnTo: `${location.pathname}${location.search}`,
-      },
-    });
-  };
-
-  return (
-    <aside className="vote-song-picker" aria-label="Choose songs">
-      <div className="vote-panel-heading">
-        <p className="vote-panel-kicker">Song window</p>
-        <button className="vote-add-songs-link" type="button" onClick={handleOpenPlaylist}>
-          Add songs
-        </button>
-      </div>
-
-      {songs.length ? (
-        <>
-          <input
-            className="vote-song-search"
-            type="search"
-            value={searchTerm}
-            onChange={(event) => setSearchTerm(event.target.value)}
-            placeholder="Search songs"
-            aria-label="Search songs by title"
-          />
-
-          {filteredSongs.length ? (
-            <div className="vote-song-list">
-              {filteredSongs.map((song) => (
-                <button
-                  className="vote-song-option"
-                  type="button"
-                  onClick={() => onAddSong(song)}
-                  key={song.id}
-                >
-                  <span>{song.name}</span>
-                  <small>
-                    {song.artist} / {song.source}
-                  </small>
-                </button>
-              ))}
-            </div>
-          ) : (
-            <p className="vote-song-empty">No matching songs</p>
-          )}
-        </>
-      ) : (
-        <p className="vote-song-empty">No songs available</p>
-      )}
-    </aside>
-  );
-}
-
-function CrabLane({ hostName, users }) {
-  const [crabIcons, setCrabIcons] = useState({});
-
-  useEffect(() => {
-    let isActive = true;
-
-    Promise.all(
-      users.map(async (user) => {
-        const hatSource = getHatSourceForCrab(user.crab, hatImages);
-        const crabIcon = await createCrabIcon(UserCrabIcon, user.crab.color, hatSource);
-
-        return [user.id, crabIcon];
-      })
-    ).then((nextCrabIcons) => {
-      if (isActive) {
-        setCrabIcons(Object.fromEntries(nextCrabIcons));
-      }
-    });
-
-    return () => {
-      isActive = false;
-    };
-  }, [users]);
-
-  return (
-    <div className="vote-crab-lane" aria-label="Players">
-      {users.map((user, index) => {
-        const isHost = user.id === hostName || user.name === hostName;
-
-        return (
-          <div
-            className="vote-player-crab"
-            style={{ animationDelay: `${index * 0.18}s` }}
-            key={user.id}
-          >
-            <img src={crabIcons[user.id] || UserCrabIcon} alt="" aria-hidden="true" />
-            <div
-              className={`vote-player-card${isHost ? " vote-player-card--host" : ""}`}
-              aria-label={isHost ? `${user.name}, host` : user.name}
-            >
-              {isHost && (
-                <img className="vote-host-crown" src={CrownIcon} alt="" aria-hidden="true" />
-              )}
-              <span>{user.name}</span>
-            </div>
-          </div>
-        );
-      })}
-    </div>
-  );
-}
 
 function VoteMovingBox({
   accountUsername,
