+++ packages/react-frontend/src/pages/game/constants.js
@@ -1,0 +1,29 @@
+export const DEFAULT_USERS = [
+  { id: "captain", name: "Captain" },
+  { id: "sailor", name: "Sailor" },
+  { id: "dj", name: "DJ" },
+  { id: "guest", name: "Guest" },
+];
+
+export const ROUND_SECONDS = 120;
+export const MIN_SCORE = -20;
+export const MAX_SCORE = 40;
+export const STACK_CARD_HEIGHT = 86;
+export const STACK_CARD_GAP = 30;
+export const VOTE_ARENA_MIN_HEIGHT = 500;
+export const VOTE_ARENA_MAX_HEIGHT = 745;
+export const VOTE_ARENA_CHROME_HEIGHT = 132;
+export const VOTE_PLAYBACK_ARENA_HEIGHT = 640;
+
+export const VOTE_BOX_COLORS = [
+  { start: "#fff4cc", middle: "#ffd67a", end: "#f2a84a" },
+  { start: "#d9f7ff", middle: "#8edff2", end: "#4db4d7" },
+  { start: "#e7f7d5", middle: "#addf76", end: "#73b95c" },
+  { start: "#ffe2ec", middle: "#f5a0bd", end: "#d86f9b" },
+  { start: "#e9e1ff", middle: "#bca3ee", end: "#8b75d6" },
+  { start: "#ffe8d3", middle: "#f5b174", end: "#dc7e45" },
+  { start: "#daf5ea", middle: "#8ed6ba", end: "#4aa889" },
+  { start: "#e3edff", middle: "#9ebdf1", end: "#638bd8" },
+  { start: "#f7e6bd", middle: "#e2c464", end: "#b99b3b" },
+  { start: "#f0dcff", middle: "#d09cf1", end: "#aa6ed0" },
+];
