+++ packages/react-frontend/src/pages/game/styles/crabLane.css
@@ -1,0 +1,95 @@
+.vote-crab-lane {
+  min-height: 230px;
+  max-height: 360px;
+  display: flex;
+  flex-wrap: wrap;
+  align-items: flex-start;
+  align-content: flex-start;
+  justify-content: center;
+  column-gap: 10px;
+  row-gap: 0;
+  overflow-y: auto;
+  overflow-x: hidden;
+  padding: 14px 6px 18px;
+}
+
+.vote-player-crab {
+  flex: 0 0 96px;
+  min-width: 96px;
+  display: grid;
+  justify-items: center;
+  gap: 0;
+  animation: crabBob 1.9s ease-in-out infinite;
+}
+
+.vote-player-crab:nth-child(even) {
+  margin-top: 54px;
+}
+
+.vote-player-crab img {
+  width: 94px;
+  height: 94px;
+  object-fit: contain;
+  position: relative;
+  z-index: 1;
+  margin-bottom: -8px;
+}
+
+.vote-player-card {
+  box-sizing: border-box;
+  width: fit-content;
+  min-width: 116px;
+  max-width: 170px;
+  min-height: 48px;
+  display: grid;
+  place-items: center;
+  border-radius: 16px;
+  background: url("../../../assets/leaflet.png") center / cover no-repeat;
+  padding: 12px 14px 9px;
+}
+
+.vote-player-card--host {
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  gap: 0.25em;
+}
+
+.vote-player-card .vote-host-crown {
+  width: 28px;
+  height: 28px;
+  display: block;
+  min-width: 16px;
+  position: static;
+  z-index: auto;
+  margin-bottom: 0;
+  object-fit: contain;
+}
+
+.vote-player-card span {
+  min-width: 88px;
+  max-width: 132px;
+  color: #1d1d1d;
+  font-size: 13px;
+  font-weight: 700;
+  line-height: 1.1;
+  overflow-wrap: anywhere;
+  white-space: normal;
+  text-align: center;
+}
+
+.vote-player-card--host span {
+  min-width: 68px;
+  max-width: 108px;
+}
+
+@keyframes crabBob {
+  0%,
+  100% {
+    transform: translateY(0);
+  }
+
+  50% {
+    transform: translateY(-14px);
+  }
+}
