+++ packages/react-frontend/src/pages/game/styles/layout.css
@@ -1,0 +1,37 @@
+.vote-game-shell {
+  min-height: 620px;
+  position: relative;
+  z-index: 1;
+  display: grid;
+  grid-template-columns: minmax(240px, 300px) minmax(0, 1fr);
+  gap: 16px;
+  color: #e7f4ff;
+}
+
+.vote-song-picker,
+.vote-arena {
+  min-width: 0;
+  border: 2px solid rgba(231, 244, 255, 0.32);
+  border-radius: 8px;
+  background: rgba(47, 80, 189, 0.62);
+}
+
+.vote-play-area {
+  min-width: 0;
+  display: grid;
+  grid-template-rows: auto auto;
+  gap: 10px;
+}
+
+.vote-panel-heading {
+  margin-bottom: 16px;
+}
+
+.vote-panel-kicker {
+  margin: 0;
+  color: #c8e6ff;
+  font-size: 12px;
+  font-weight: 700;
+  letter-spacing: 0.08em;
+  text-transform: uppercase;
+}
