+++ packages/react-frontend/src/pages/game/styles/songPicker.css
@@ -1,0 +1,118 @@
+.vote-song-picker {
+  height: 720px;
+  max-height: 775px;
+  display: flex;
+  flex-direction: column;
+  padding: 16px;
+  align-self: start;
+  overflow: hidden;
+  background: rgba(8, 48, 71, 0.86);
+  box-shadow: 0 8px 22px rgba(0, 0, 0, 0.28);
+}
+
+.vote-add-songs-link {
+  display: inline-flex;
+  align-items: center;
+  justify-content: center;
+  min-height: 42px;
+  border: 2px solid rgba(231, 244, 255, 0.58);
+  border-radius: 999px;
+  background: #2f50bd;
+  color: #ffffff;
+  cursor: pointer;
+  font: inherit;
+  font-size: 20px;
+  font-weight: 700;
+  margin: 6px 0 0;
+  padding: 8px 18px;
+}
+
+.vote-add-songs-link:hover,
+.vote-add-songs-link:focus-visible {
+  border-color: #f7b244;
+  outline: none;
+}
+
+.vote-song-search {
+  box-sizing: border-box;
+  width: 100%;
+  flex: 0 0 auto;
+  border: 2px solid rgba(231, 244, 255, 0.58);
+  border-radius: 8px;
+  background: #eefbff;
+  color: #083047;
+  font: inherit;
+  margin-bottom: 10px;
+  padding: 12px;
+}
+
+.vote-song-search:focus {
+  border-color: #f7b244;
+  outline: none;
+}
+
+.vote-song-search::placeholder {
+  color: #2f50bd;
+}
+
+.vote-song-list {
+  flex: 1 1 auto;
+  min-height: 0;
+  display: flex;
+  flex-direction: column;
+  gap: 10px;
+  overflow-y: auto;
+  padding-right: 4px;
+}
+
+.vote-song-empty {
+  min-height: 160px;
+  display: grid;
+  place-items: center;
+  border: 2px dashed rgba(231, 244, 255, 0.42);
+  border-radius: 8px;
+  font-size: 18px;
+  margin: 0;
+  padding: 20px;
+  text-align: center;
+}
+
+.vote-song-option {
+  min-height: 90px;
+  display: flex;
+  flex-direction: column;
+  align-items: flex-start;
+  justify-content: center;
+  gap: 4px;
+  border: 2px solid rgba(231, 244, 255, 0.58);
+  border-radius: 8px;
+  background: #eefbff;
+  color: #ffffff;
+  cursor: pointer;
+  font: inherit;
+  padding: 18px 12px;
+  text-align: left;
+  white-space: normal;
+}
+
+.vote-song-option:hover,
+.vote-song-option:focus-visible {
+  border-color: #f7b244;
+  background: #ffffff;
+}
+
+.vote-song-option span {
+  color: #083047;
+  font-size: 16px;
+  font-weight: 700;
+  line-height: 1.2;
+  overflow-wrap: anywhere;
+}
+
+.vote-song-option small {
+  color: #2f50bd;
+  font-size: 13px;
+  font-weight: 700;
+  line-height: 1.2;
+  overflow-wrap: anywhere;
+}
