+++ packages/react-frontend/src/pages/game/youtubeApi.js
@@ -1,0 +1,36 @@
+import { getSongVideoId } from "../Playlist/playlistStorage";
+
+let youtubeApiPromise;
+
+export function loadYouTubeApi() {
+  if (typeof window === "undefined") {
+    return Promise.resolve(null);
+  }
+
+  if (window.YT?.Player) {
+    return Promise.resolve(window.YT);
+  }
+
+  if (!youtubeApiPromise) {
+    youtubeApiPromise = new Promise((resolve) => {
+      const previousReady = window.onYouTubeIframeAPIReady;
+
+      window.onYouTubeIframeAPIReady = () => {
+        previousReady?.();
+        resolve(window.YT);
+      };
+
+      if (!document.querySelector('script[src="https://www.youtube.com/iframe_api"]')) {
+        const script = document.createElement("script");
+        script.src = "https://www.youtube.com/iframe_api";
+        document.body.appendChild(script);
+      }
+    });
+  }
+
+  return youtubeApiPromise;
+}
+
+export function getPlayableVideoId(song) {
+  return getSongVideoId(song) || getSongVideoId({ songLink: song?.songLink });
+}
